adds sidebar on mobile
