+++ public/index.html
@@ -8,7 +8,6 @@
       rel="stylesheet"
     />
     <title>Melissa Pastore | Software Engineer</title>
-    <script defer src="/bundle.js"></script>
   </head>
   <body>
     <div id="root"></div>
