+++ src/Components/Navbar.js
@@ -1,12 +1,26 @@
-import React from "react";
+import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import Sidebar from "./Sidebar";
+import IconButton from "@material-ui/core/IconButton";
+import MenuIcon from "@material-ui/icons/Menu";
 
 const Navbar = () => {
+  const [open, setOpen] = useState(false);
+
+  const handleDrawerOpen = () => {
+    console.log("Handle open");
+    setOpen(true);
+  };
+
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
+
   return (
     <nav id="navbar">
-      <h1>Melissa Pastore</h1>
-      <Sidebar />
+      <h1 id="navbar-heading">Melissa Pastore</h1>
+
+      <Sidebar handleDrawerClose={handleDrawerClose} open={open} />
       <Link className="link" to="/">
         About
       </Link>
@@ -19,6 +33,16 @@ const Navbar = () => {
       <Link className="link" to="/resume">
         Experience
       </Link>
+      <div id="menu-button">
+        <IconButton
+          onClick={handleDrawerOpen}
+          color="inherit"
+          aria-label="open drawer"
+          edge="start"
+        >
+          <MenuIcon />
+        </IconButton>
+      </div>
     </nav>
   );
 };
